Further separated logger, util, recipebases
