+++ pythonforandroid/util.py
@@ -1,0 +1,44 @@
+import contextlib
+from os.path import exists
+try:
+    from urllib.request import FancyURLopener
+except ImportError:
+    from urllib import FancyURLopener
+
+
+class ChromeDownloader(FancyURLopener):
+    version = (
+        'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 '
+        '(KHTML, like Gecko) Chrome/28.0.1500.71 Safari/537.36')
+
+urlretrieve = ChromeDownloader().retrieve
+
+
+@contextlib.contextmanager
+def current_directory(new_dir):
+    cur_dir = getcwd()
+    logger.info(''.join((Err_Fore.CYAN, '-> directory context ', new_dir,
+                         Err_Fore.RESET)))
+    chdir(new_dir)
+    yield
+    logger.info(''.join((Err_Fore.CYAN, '<- directory context ', cur_dir,
+                         Err_Fore.RESET)))
+    chdir(cur_dir)
+
+
+@contextlib.contextmanager
+def temp_directory():
+    temp_dir = mkdtemp()
+    try:
+        logger.debug(''.join((Err_Fore.CYAN, ' + temp directory used ',
+                              temp_dir, Err_Fore.RESET)))
+        yield temp_dir
+    finally:
+        shutil.rmtree(temp_dir)
+        logger.debug(''.join((Err_Fore.CYAN, ' - temp directory deleted ',
+                              temp_dir, Err_Fore.RESET)))
+
+
+def ensure_dir(filename):
+    if not exists(filename):
+        makedirs(filename)
